+++ Source/Classes/ObjectParser.ps1
@@ -1,228 +1,212 @@
 <#
-# Object Parser
+.SYNOPSIS
+    Class to support Object Graph Tools
+.DESCRIPTION
+    This class provides general properties and method to recursively
+    iterate through to PowerShell Object Graph nodes.
 
-This class provides general properties and method to recursively
-iterate through to PowerShell Object Graph nodes.
+    For details, see:
 
-## Example "Iterate trough each recursive node and return it property path"
+    * [PowerShell Object Parser][1] for details on the `[PSNode]` properties and methods.
+    * [Extended-Dot-Notation][2] for details on path selectors.
 
-The following function recursively iterates through all the property nodes (`PSNodes`)
-of an object-graph and returns the path to each object.
+.LINK
+    [1]: https://github.com/iRon7/ObjectGraphTools/blob/main/Docs/ObjectParser.md "PowerShell Object Parser"
+    [2]: https://github.com/iRon7/ObjectGraphTools/blob/main/Docs/Xdn.md "Extended Dot Notation"
+#>
 
-```PowerShell
-function Iterate([PSNode]$Node) { # Basic iterator
-    $Node.PathName
-    if ($Node -is [PSCollectionNode]) {
-        $Node.ChildNodes.foreach{ Iterate $_ }
+using namespace System.Collections.Generic
+using namespace System.Management.Automation
+using namespace System.Management.Automation.Language
+
+enum XdnPredecessor { Root; Parent; Ancestor }
+enum XdnType { Root; Ancestor; Index; Child; Descendant }
+
+class Literal {
+    hidden [String]$_String
+
+    Literal([String]$String) {
+        $this._String = $String
+    }
+
+    [String] ToString() {
+        return $this._String
+    }
+
+    [String] Quoted() {
+        return "'" + $this._String.Replace("'", "''") + "'"
     }
 }
 
-$Object = $Json | ConvertFrom-Json
-$PSNode = [PSNode]::ParseInput($Object)
-Iterate $PSNode
-```
+class XdnPath {
+    hidden static $_PSReadLineOption
+    hidden static $Verbatim = '^[\?\*\p{L}\p{Lt}\p{Lm}\p{Lo}_][\?\*\p{L}\p{Lt}\p{Lm}\p{Lo}\p{Nd}_]*$' # https://stackoverflow.com/questions/62754771/unquoted-key-rules-and-best-practices
+
+    hidden $_Entries = [List[KeyValuePair[XdnType, Object]]]::new()
+
+    hidden [Object]get_Entries() { return ,$this._Entries }
+
+    Add ([XdnType]$XdnType, $Value) {
+        $KeyValuePair = [KeyValuePair[XdnType, Object]]::new($XdnType, $Value)
+        $this._Entries.Add($KeyValuePair)
+    }
+
+    XdnPath ([String]$Path) {
+        if (-not $this._Entries.Count -and $Path -notmatch '^(?<=([^`]|^)(``)*)[\.]') { $this.Add('Root', $Null) }
+        $Length = [Int]::MaxValue
+        [XdnPredecessor]$Predecessor = 'Root'
+        while ($Path) {
+            if ($Path.Length -ge $Length) { break }
+            $Length = $Path.Length
+            if ($Path[0] -in "'", '"') {
+                $Ast = [Parser]::ParseInput($Path, [ref]$Null, [ref]$Null)
+                $StringAst = $Ast.EndBlock.Statements.PipelineElements.Find({ $args[0] -is [StringConstantExpressionAst] }, $false)[0]
+                if ($Null -ne $StringAst) {
+                    if ($Predecessor -eq 'Ancestor') { $this.Add('Descendant', [Literal]$StringAst.Value) }
+                    else { $this.Add('Child', [Literal]$StringAst.Value) }
+                    $Path = $Path.SubString($StringAst.Extent.EndOffset)
+                }
+                else { # Likely a quoting error
+                    if ($Predecessor -eq 'Ancestor') { $this.Add('Descendant', [Literal]$Path) }
+                    else { $this.Add('Child', [Literal]$Path) }
+                    $Path = $Null
+                }
+            }
+            else {
+                $Match = [regex]::Match($Path, '(?<=([^`]|^)(``)*)[\.\[\-]')
+                if ($Match.Success -and $Match.Index -eq 0) {
+                    $IndexEnd  = if ($Match.Value -eq '[') { $Path.IndexOf(']') }
+                    $Ancestors = if ($Match.Value -eq '.' -and $Path -Match '^\.\.+') { $Matches[0].Length - 1 }
+                    if ($IndexEnd -gt 0 -and $Predecessor -ne 'Ancestor') {
+                        $Index = $Path.SubString(1, ($IndexEnd - 1))
+                        $CommandAst = [Parser]::ParseInput($Index, [ref]$Null, [ref]$Null).EndBlock.Statements.PipelineElements
+                        if ($CommandAst -is [CommandExpressionAst]) { $Index = $CommandAst.expression.Value }
+                        $this.Add('Index', $Index)
+                        $Path = $Path.SubString(($IndexEnd + 1))
+                    }
+                    elseif ($Ancestors) {
+                        $Predecessor = 'Parent'
+                        $this.Add('Ancestor', $Ancestors)
+                        $Path = $Path.Substring($Ancestors + 1)
+                    }
+                    elseif ($Match.Value -eq '.') {
+                        $Predecessor = 'Parent'
+                        $Path = $Path.Substring(1)
+                    }
+                    elseif ($Match.Value -eq '-' -and ($this._Entries.Count -and $this._Entries[-1].Key -ne 'Descendant')) {
+                        $Predecessor = 'Ancestor'
+                        $Path = $Path.Substring(1)
+                    }
+                    else { # Likely a selector error
+                        if ($Predecessor -eq 'Ancestor') { $this.Add('Descendant', $Match.Value) } else { $this.Add('Child', $Match.Value) }
+                        $Path = $Path.Substring(1)
+                    }
+                }
+                elseif ($Match.Success) {
+                    $Name = $Path.SubString(0, $Match.Index)
+                    if ($Predecessor -eq 'Ancestor') { $this.Add('Descendant', $Name) } else { $this.Add('Child', $Name) }
+                    $Path = $Path.SubString($Match.Index)
+                }
+                else {
+                    if ($Predecessor -eq 'Ancestor') { $this.Add('Descendant',$Path) } else { $this.Add('Child', $Path) }
+                    $Path = $Null
+                }
+            }
+        }
+    }
+    [String] ToString([String]$VariableName, [Bool]$Color) {
+        if ($Null -eq [XdnPath]::_PSReadLineOption) {
+            if (Get-Command Get-PSReadLineOption -ErrorAction SilentlyContinue) {
+                [XdnPath]::_PSReadLineOption = Get-PSReadLineOption
+            }
+            else { [XdnPath]::_PSReadLineOption = $false }
+        }
+
+        $Option = [XdnPath]::_PSReadLineOption
+        if (-not $Option) { $Color = $false }
+        $RegularColor  = if ($Color) { $Option.VariableColor }
+        $WildcardColor = if ($Color) { $Option.EmphasisColor }
+        $SpecialColor  = if ($Color) { $option.CommandColor }
+        $ErrorColor    = if ($Color) { $Option.ErrorColor }
+        $ResetColor    = if ($Color) { [char]0x1b + '[39m' }
+
+        $Path = [System.Text.StringBuilder]::new()
+        $PreviousEntry = $Null
+        foreach ($Entry in $this._Entries) {
+            $Value = $Entry.Value
+            $Append = Switch ($Entry.Key) {
+                Root {
+                    "$SpecialColor$VariableName$ResetColor"
+                }
+                Ancestor {
+                    "$SpecialColor$('.' * $Value)$ResetColor"
+                }
+                Index {
+                    $Dot = if (-not $PreviousEntry -or $PreviousEntry.Key -eq 'Ancestor') { "$SpecialColor." }
+                    if ([int]::TryParse($Value, [Ref]$Null)) { "$Dot$RegularColor[$Value]$ResetColor" }
+                    else { "$ErrorColor[$Value]$ResetColor" }
+                }
+                Child {
+                    if ($Value -is [Literal])                     { "$RegularColor.$($Value.Quoted())$ResetColor" }
+                    elseif ($Value -NotMatch [XdnPath]::Verbatim) { "$ErrorColor.$Value$ResetColor" }
+                    elseif ($Value -Match '[\?\*]')               { "$WildcardColor.$Value$ResetColor" }
+                    else                                          { "$RegularColor.$Value$ResetColor" }
+                }
+                Descendant {
+                    if ($Value -is [Literal])                     { "$SpecialColor-$ErrorColor$($Value.Quoted())$ResetColor" }
+                    elseif ($Value -NotMatch [XdnPath]::Verbatim) { "$SpecialColor-$ErrorColor$Value$ResetColor" }
+                    elseif ($Value -Match '[\?\*]')               { "$SpecialColor-$WildcardColor$Value$ResetColor" }
+                    else                                          { "$SpecialColor-$RegularColor$Value$ResetColor" }
+                }
+            }
+            $Path.Append($Append)
+            $PreviousEntry = $Entry
+        }
+        return $Path.ToString()
+    }
+    [String] ToString()                           { return $this.ToString($Null        , $false)}
+    [String] ToString([String]$VariableName)      { return $this.ToString($VariableName, $false)}
+    [String] ToColorString()                      { return $this.ToString($Null,         $true)}
+    [String] ToColorString([String]$VariableName) { return $this.ToString($VariableName, $true)}
+
+    static XdnPath() { # https://stackoverflow.com/questions/77752014/how-to-type-convert-a-derived-class
+        $FormatData = @'
+        <Configuration>
+        <ViewDefinitions>
+            <View>
+            <Name>XdnPath</Name>
+            <OutOfBand />
+            <ViewSelectedBy>
+                <TypeName>XdnPath</TypeName>
+            </ViewSelectedBy>
+                <CustomControl>
+                <CustomEntries>
+                    <CustomEntry>
+                    <CustomItem>
+                        <ExpressionBinding>
+                        <ScriptBlock>
+                            <![CDATA[$_.ToColorString('<Root>')]]>
+                        </ScriptBlock>
+                        </ExpressionBinding>
+                    </CustomItem>
+                    </CustomEntry>
+                </CustomEntries>
+                </CustomControl>
+            </View>
+            </ViewDefinitions>
+        </Configuration>
+'@
+        $TempFile = [IO.Path]::GetTempPath() + "XdnPath.ps1xml"
+        Out-File -InputObject $FormatData -LiteralPath $TempFile -Encoding ASCII
+        Update-FormatData -PrependPath $TempFile
+    }
+}
+
+enum PSNodeOrigin { Root; List; Map }
 
-## Class hierarchy
-
-The general class is called `[PSNode]`and has a hierarchy of sub-classes
-
-```plaintext
-         [PSLeafNode]
-        /
-[PSNode]                    [PSListNode]
-        \                  /
-         [PSCollectionNode]             [PSDictionaryNode]
-                           \           /
-                            [PSMapNode]
-                                       \
-                                        [PSObjectNode]
-```
-
-### `[PSNode]`
-
-The base type of a PSNode. Each PSNode will have at least 1 additional PSNode derivative
-listed below.
-
-### `[PSLeafNode]`
-
-A PSLeafNode terminates a PSNode branch and doesn't have any child nodes attached.
-A value embedded by a PSLeafNode is not enumerable and doesn't contain any object properties.
-
-### `[PSCollectionNode]`
-
-A PSCollectionNode represents a PSNode containing a collection child nodes.
-A value embedded by a PSCollectionNode is enumerable or contains object properties.
-
-### `[PSListNode]`
-
-A PSListNode represents a  PSNode listing any number (or none) child nodes.
-A value embedded by a PSListNode supports the `IList` interface but excludes any value that.
-support the IDictionary interface.
-
-### `[PSMapNode]`
-
-A PSMapNode represents a PSNode containing any number (or none) child nodes.
-A value embedded by a PSMapNode supports the IDictionary interface or contains object properties.
-
-### `[PSDictionaryNode]`
-
-A PSDictionaryNode represents a PSNode containing any number (or none) child nodes.
-A value embedded by a PSDictionaryNode supports the IDictionary interface.
-
-### `[PSObjectNode]`
-
-A PSObjectNode represents a PSNode containing any number (or none) child nodes.
-A value embedded by a PSObjectNode contains object properties meaning that it is either of type
-`[PSCustomObject]` or `[ComponentModel.Component]`.
-
-## Constructors
-
-There are no noteworthy constructors.
-To create a new PSNode instance, use the static `[PSNode]::ParseInput(<Object-Graph>)]`
-method or the `ChildNodes` property or `GetChildNode(<name>)` of an existing PSNode
-instance
-
-## Properties
-
-### `Name` (ReadOnly)
-
-The name of the embedded property defined by its parent.
-The name is `$Null` if the embedded node is the root node.
-
-### `Value`
-
-The actual object, item, property or value embedded by the PSNode.
-The value might be modified but should be of the same structure (`[PSLeafNode]`, `[PSListNode]`,
-`[PSDictionaryNode]` or `[PSObjectNode]`) as the original node type.
-
-### `Depth` (ReadOnly)
-
-The depth where the current PSNode resides in the PSNode hierarchy (aka tree).
-An error will occur if the depth exceed the `MaxDepth` setting.
-
-### `MaxDepth`
-
-The maximum iteration depth of the embedded graph object.
-The `MaxDepth` value can be read at every level but can only be set on the root node:
-`[PSNode].RootNode.MaxDepth = <Maximum Depth>`
-
-### `ValueType` (ReadOnly)
-
-The type of the value embedded by the PSNode.
-`$Null` if the embedded value is `$Null`
-
-### `NodeOrigin` (ReadOnly)
-
-Defines whether the parent node is a `[PSListNode]` type, a `[PSMapNode]` type.
-The NodeOrigin is `Root` if the current node has no parent node.
-
-### `ParentNode` (ReadOnly)
-
-Refers to node containing the current PSNode and possible siblings.
-
-### `RootNode` (ReadOnly)
-
-Refers to the top node containing the current PSNode and all its decedents.
-
-### `ChildNodes` (ReadOnly) (`[PSCollectionNodes]` only)
-
-Returns all the child nodes contained by the current PSNode.
-To retrieve a specific child node, use the `GetChildNode(<Name>)` method.
-
-### `DescendantNodes` (ReadOnly) (`[PSCollectionNodes]` only)
-
-Returns all the descendant nodes (up and till the `$MaxDepth` level) contained
-by the current PSNode.
-
-### `Count` (ReadOnly) (`[PSCollectionNodes]` only)
-
-Returns the number of items or properties contained by the embedded value.
-This number is equal to the number of child nodes contained by the current node.
-
-### `Names` (ReadOnly) (`[PSCollectionNodes]` only)
-
-Returns the property or key names of the embedded value. If the PSNode is of
-`[PSListNode]` type list a indices (starting from zero) is returned.
-The name of each child node is equal to the name (or index) that identifies item
-of the embedded value
-
-### `Values` (ReadOnly) (`[PSCollectionNodes]` only)
-
-Returns all the value of the items or properties of the embedded value.
-A PSNode derived from a `[PSLeafNode]` type doesn't have a `Values` property.
-
-### `Path` (ReadOnly)
-
-Returns all the branch of PSNodes starting from the root PSNode up and till
-the current PSNode
-
-### `PathName` (ReadOnly)
-
-Returns an unique path (string) identifying the current node in the PSNode tree
-starting from the root node.
-The path might be used to directly target the property or item of a object-graph
-aka the value contained by the root node.
-
-## Methods
-
-### `[PSNode]::ParseInput(<object-graph> [, <maximal depth = 10>])`
-
-This static method converts a object-graph to a [PSNode] structure and supplies access
-to the underlying child nodes.
-The `<maximal depth>` argument, set the maximal depth of the properties  that will be
-recursively retrieve. When the maximal depth is reached, an error is throw.
-The default maximal depth is defined by the static property `[PSNode]::MaxDepth` (default: 10)
-
-### `GetChildNode(<name>)` (`[PSCollectionNodes]` only)
-
-Returns a specific child node (`[PSNode]`) selected by the name (or index) of the embedded
-
-> [!Note]
-> The `GetChildNode` has a Shorthand ("alias"): `_(<name>)` which shouldn't be used
-> in scripts as the its name or functionality might change in the future
-
-### `GetDescendentNode(<path>)` (`[PSCollectionNodes]` only)
-
-Returns a specific child node (`[PSNode]`) selected by the path of the embedded object
-The path might be either:
-
-* As [String] a "dot-property" selection as defined by the `PathName` property a specific node.
-* A array of strings (dictionary keys or Property names) and/or Integers (list indices).
-* A object (`PSNode[]`) list where each `Name` property defines the path
-
-> [!Note]
-> The `GetDescendentNode` has a Shorthand ("alias"): `Get(<name>)` which shouldn't be used
-> in scripts as the its name or functionality might change in the future
-
-### `GetDescendentNodes(<generations>)` (`[PSCollectionNodes]` only)
-
-Returns all descendant nodes of the current node for a the defined number of generations.
-
-> [!Note]
-> The number of generations will not surpass the `MaxDepth` defined at the root.
-
-### `GetItem(<name>)` (`[PSCollectionNodes]` only)
-
-Returns the value of a specific item identified by `<name>` of the embedded collection
-or object.
-
-### `SetItem(<name>, <value>)` (`[PSCollectionNodes]` only)
-
-Sets the value of a specific item identified by `<name>` of the embedded collection
-or object. The new value should be of the same structure (`[PSLeafNode]`, `[PSListNode]`,
-`[PSDictionaryNode]` or `[PSObjectNode]`) as the original node type.
-
-### `Contains(<name>)` (`[PSCollectionNodes]` only)
-
-Determines whether a specific item identified by `<name>` is contained by th embedded
-collection or object.
-#>
-
-Using NameSpace System.Management.Automation.Language
-enum PSNodeOrigin { Root; List; Map}
 
 Class PSNode {
-    static [int]$DefaultMaxDepth = 10
+    static [int]$DefaultMaxDepth = 20
     hidden $_Name
     [Int]$Depth
     hidden $_Value
@@ -257,7 +241,7 @@ Class PSNode {
     }
 
     hidden set_MaxDepth($MaxDepth) {
-        if ($this.NodeOrigin -eq 'Root') {
+        if (-not $this.ChildType) {
             $this._MaxDepth = $MaxDepth
         }
         else {
@@ -272,7 +256,7 @@ Class PSNode {
         else { return $this._Value.getType() }
     }
 
-    hidden static [String]getPSNodeType($Object) {
+    hidden static [String]GetPSNodeType($Object) {
             if ($Object -is [Management.Automation.PSCustomObject]) { return 'PSObjectNode' }
         elseif ($Object -is [ComponentModel.Component])             { return 'PSObjectNode' }
         elseif ($Object -is [Collections.IDictionary])              { return 'PSDictionaryNode' }
@@ -300,13 +284,14 @@ Class PSNode {
 
     hidden [PSNode] Append($Object) {
         $Node = [PSNode]::ParseInput($Object)
-        $Node.Depth      = $this.Depth + 1
-        $Node.RootNode   = $this.RootNode
-        $Node.ParentNode = $this
+        $Node.Depth       = $this.Depth + 1
+        $Node.RootNode    = $this.RootNode
+        $Node.ParentNode  = $this
+        $Node._NodeOrigin = if ($this -is [PSListNode]) { 'List' } elseif ($this -is [PSMapNode]) { 'Map' }
         return $Node
     }
 
-    hidden [System.Collections.Generic.List[PSNode]] get_Path() {
+    hidden [List[PSNode]] get_Path() {
         if ($this._Path.Count -eq 0) {
             if ($this.ParentNode) { $ParentPath = $this.ParentNode.get_Path() } else { $ParentPath =  @() }
             $this._Path = $ParentPath + $this # This will shallow copy the parent path
@@ -314,22 +299,93 @@ Class PSNode {
         return $this._Path
     }
 
-    hidden [String] get_PathName() {
+    [String] GetPathName() {
         if ($Null -eq $this._PathName) {
-            $ParentPathName = if ($this.ParentNode) { $this.ParentNode.get_PathName() }
+            $ParentPathName = if ($this.ParentNode) { $this.ParentNode.GetPathName() }
             $Name =
                 if ($this._NodeOrigin -eq 'List') {
                     "[$($this._Name)]"
                 }
                 elseif ($this._NodeOrigin -eq 'Map') {
-                    if     ($this.Name -is [ValueType])        { ".$($this._Name)" }
-                    elseif ($this.Name -isnot [String])        { ".[$($this._Name.GetType())]'$($this._Name)'" }
-                    elseif ($this.Name -Match '^[_,a-z]+\w*$') { ".$($this._Name)" }
-                    else                                       { ".'$($this._Name)'" }
+                    $Dot = if ($this.ParentNode.ParentNode) { '.' }
+                    if     ($this.Name -is [ValueType])        { "$Dot$($this._Name)" }
+                    elseif ($this.Name -isnot [String])        { "$Dot[$($this._Name.GetType())]'$($this._Name)'" }
+                    elseif ($this.Name -Match '^[_,a-z]+\w*$') { "$Dot$($this._Name)" }
+                    else                                       { "$Dot'$($this._Name)'" }
                 }
             $this._PathName = $ParentPathName + $Name
         }
         return $this._PathName
+    }
+
+    [String] GetPathName($VariableName) {
+        $PathName = $this.GetPathName()
+        if ($PathName -and $PathName[0] -in '.', '-', '[' ) {
+            return "$VariableName$PathName"
+        }
+        else {
+            return "$VariableName.$PathName"
+        }
+    }
+    hidden [String] get_PathName() { return $this.GetPathName() }
+
+    hidden CollectNodes($PathNodes, [XdnPath]$Path, [Int]$PathIndex) {
+        $Entry = $Path._Entries[$PathIndex]
+        $NextIndex = if ($PathIndex -lt $Path._Entries.Count -1) { $PathIndex + 1 }
+        switch ($Entry.Key) {
+            Root {
+                $Node = $this.RootNode
+                if ($NextIndex) { $Node.CollectNodes($PathNodes, $Path, $NextIndex) }
+                else { $PathNodes[$Node.get_PathName()] = $Node }
+            }
+            Ancestor {
+                $Node = $this
+                for($i = $Entry.Value; $i -gt 0 -and $Node.ParentNode; $i--) { $Node = $Node.ParentNode }
+                if ($i -eq 0) { # else: reached root boundary
+                    if ($NextIndex) { $Node.CollectNodes.GetNode($PathNodes, $Path, $NextIndex) }
+                    else { $PathNodes[$Node.get_PathName()] = $Node }
+                }
+            }
+            Index {
+                if ($this -is [PSListNode] -and [Int]::TryParse($Entry.Value, [Ref]$Null)) {
+                    $Node = $this.GetChildNode([Int]$Entry.Value)
+                    if ($NextIndex) { $Node.CollectNodes($PathNodes, $Path, $NextIndex) }
+                    else { $PathNodes[$Node.get_PathName()] = $Node }
+                }
+            }
+            Default { # Child, Descendant
+                if ($this -is [PSListNode]) { # Member access enumeration
+                    foreach ($Node in $this.get_ChildNodes()) {
+                        $Node.CollectNodes($PathNodes, $Path, $PathIndex)
+                    }
+                }
+                elseif ($this -is [PSMapNode]) {
+                    $Found = $False
+                    $ChildNodes = $this.get_ChildNodes()
+                    foreach ($Node in $ChildNodes) {
+                        $Exists = if ($Entry.Value -is [Literal]) { $Node.Name -eq $Entry.Value } else { $Node.Name -like $Entry.Value }
+                        if ($Exists) {
+                            $Found = $True
+                            if ($NextIndex) { $Node.CollectNodes($PathNodes, $Path, $NextIndex) }
+                            else { $PathNodes[$Node.get_PathName()] = $Node }
+                        }
+                    }
+                    if (-not $Found -and $Entry.Key -eq 'Descendant') {
+                        foreach ($Node in $ChildNodes) {
+                            $Node.CollectNodes($PathNodes, $Path, $PathIndex)
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    [Object] GetNode([XdnPath]$Path) {
+        $PathNodes = [system.collections.generic.dictionary[String, PSNode]]::new() # Case sensitive (case insensitive map nodes use the same name)
+        $this.CollectNodes($PathNodes, $Path, 0)
+        if ($PathNodes.Count -eq 0) { return @() }
+        if ($PathNodes.Count -eq 1) { return $PathNodes[$PathNodes.Keys] }
+        else                        { return [PSNode[]]$PathNodes.Values }
     }
 }
 
@@ -365,8 +421,8 @@ Class PSCollectionNode : PSNode {
         Write-Warning "Expected $SelectionName to be a $CollectionType selector for: <Object>$($Node.PathName)"
     }
 
-    hidden[Collections.Generic.List[Ast]] GetAstSelectors ($Ast) {
-        $List = [Collections.Generic.List[Ast]]::new()
+    hidden [List[Ast]] GetAstSelectors ($Ast) {
+        $List = [List[Ast]]::new()
         if ($Ast -isnot [Ast]) {
             $Ast = [Parser]::ParseInput("`$_$Ast", [ref]$Null, [ref]$Null)
             $Ast = $Ast.EndBlock.Statements.PipeLineElements.Expression
@@ -385,51 +441,22 @@ Class PSCollectionNode : PSNode {
         return $List
     }
 
-    [Object]GetDescendentNode($Path) {
-        if ($Path -is [String]) {
-            $String = if ($Path.StartsWith('.') -or $Path.StartsWith('[')) { "`$_$Path"} else { "`$_.$Path"}
-            $Ast  = [Parser]::ParseInput($String, [ref]$Null, [ref]$Null)
-            $Ast = $Ast.EndBlock.Statements.PipeLineElements.Expression
-            $Selectors = $this.GetAstSelectors($Ast)
-        }
-        elseif ($Path -is [PSNode]) { $Selectors = $Path.Path }
-        else { $Selectors = $Path }
-        $Node = $this
-        foreach ($Selector in $Selectors) {
-            if ($Node -is [PSLeafNode]) {
-                Throw "Can not select child node in <object>$($Node.PathName) as it is a leaf node."
-            }
-            elseif ($Selector -is [IndexExpressionAst]) {
-                $Name = $Selector.Index.Value
-                if ($Node -isnot [PSListNode]) { $this.WarnSelector($Node, $Name) }
-            }
-            elseif ($Selector -is [MemberExpressionAst]) {
-                $Name = $Selector.Member.Value
-                if ($Node -isnot [PSMapNode]) { $this.WarnSelector($Node, $Name) }
-            }
-            else {
-                $Name = if ($Selector.PSObject.Properties['Name']) { $Selector.Name } else { $Selector }
-                if ($Selector -is [PSNode]) {
-                    if ($Selector.PSNodeOrigin -eq 'List' -and  $Node -isnot [PSListNode]) { $this.WarnSelector($Node, $Name) }
-                    if ($Selector.PSNodeOrigin -eq 'Map'  -and  $Node -isnot [PSMapNode])  { $this.WarnSelector($Node, $Name) }
-                }
-                elseif ($Name -is [Int]) {
-                    if ($Node.ValueType.IsGenericType -and $Node.ValueType.GetGenericArguments()[0].Name -eq 'String') { $this.WarnSelector($Node, $Name) }
-                }
-                else {
-                    if ($Node -isnot [PSMapNode]) {$this.WarnSelector($Node, $Name) }
-                }
-            }
-            if ($Null -ne $Name) { $Node = $Node.GetChildNode($Name) } # Not sure yet whether $Null should selected the current node or the root node...
-        }
-        return $Node
+    [List[PSNode]]GetNodes() {
+        return $this.GetNodes(0, 0, $false)
     }
-
-    [Collections.Generic.List[PSNode]]GetDescendentNodes() { return $this.GetDescendentNodes(0) }
-    hidden [Object]get_ChildNodes()      { return ,[PSNode[]]@($this.GetDescendentNodes(0)) }
-    hidden [Object]get_DescendantNodes() { return ,[PSNode[]]@($this.GetDescendentNodes(-1)) }
-    hidden [Object]_($Name)              { return $this.GetChildNode($Name) }       # CLI Shorthand ("alias") for GetChildNode (don't use in scripts)
-    hidden [Object]Get($Path)            { return $this.GetDescendentNode($Path) }  # CLI Shorthand ("alias") for GetDescendentNode (don't use in scripts)
+    [List[PSNode]]GetNodes([Int]$Levels) {
+        return $this.GetNodes($Levels, 0, $false)
+    }
+    [List[PSNode]]GetNodes([PSNodeOrigin]$NodeOrigin, [Bool]$Leaf) {
+        return $this.GetNodes(0, $NodeOrigin, $Leaf)
+    }
+    hidden [Object]get_ChildNodes()      { return [PSNode[]]$this.GetNodes(0,  0,      $false) }
+    hidden [Object]get_ListChildNodes()  { return [PSNode[]]$this.GetNodes(0,  'List', $false) }
+    hidden [Object]get_MapChildNodes()   { return [PSNode[]]$this.GetNodes(0,  'Map',  $false) }
+    hidden [Object]get_DescendantNodes() { return [PSNode[]]$this.GetNodes(-1, 0,      $false) }
+    hidden [Object]get_LeafNodes()       { return [PSNode[]]$this.GetNodes(-1, 0,      $true) }
+    hidden [Object]_($Name)              { return [PSNode]$this.GetChildNode($Name) }       # CLI Shorthand ("alias") for GetChildNode (don't use in scripts)
+    # hidden [Object]Get($Path)                { return $this.GetDescendantNode($Path) }  # CLI Shorthand ("alias") for GetDescendantNode (don't use in scripts)
 }
 
 Class PSListNode : PSCollectionNode {
@@ -461,16 +488,16 @@ Class PSListNode : PSCollectionNode {
         $this._Value[$Index] = $Value
     }
 
-    [Collections.Generic.List[PSNode]]GetDescendentNodes([Int]$Levels) {
-        $List = [Collections.Generic.List[PSNode]]::new()
+    [List[PSNode]]GetNodes([Int]$Levels, [PSNodeOrigin]$NodeOrigin, [Bool]$Leaf) {
+        $List = [List[PSNode]]::new()
         if (-not $this.MaxDepthReached()) {
             for ($Index = 0; $Index -lt $this._Value.Get_Count(); $Index++) {
                 $Node = $this.Append($this._Value[$Index])
-                $Node._NodeOrigin = 'List'
                 $Node._Name = $Index
-                $List.Add($Node)
-                if ($Levels -ne 0 -and $Node -is [PSCollectionNode]) {
-                    $list.AddRange($Node.GetDescendentNodes($Levels - 1))
+                if ($NodeOrigin -in 0, 'List' -or ($Leaf -and $Node -is [PSLeafNode])) { $List.Add($Node) }
+                if ($Node -is [PSCollectionNode] -and ($Levels -ne 0 -or $NodeOrigin -eq 'Map')) { # $NodeOrigin -eq 'Map' --> Member Access Enumeration
+                    $Levels_1 = if ($Levels -gt 0) { $Levels - 1 } else { $Levels }
+                    $list.AddRange($Node.GetNodes($Levels_1, $NodeOrigin, $Leaf))
                 }
             }
         }
@@ -484,14 +511,13 @@ Class PSListNode : PSCollectionNode {
             throw "The <Object>$($this.PathName) doesn't contain a child index: $Index"
         }
         $Node = $this.Append($this._Value[$Index])
-        $Node._NodeOrigin = 'List'
         $Node._Name = $Index
         return $Node
     }
 
     [Int]GetHashCode() {
         $HashCode = '@()'.GetHashCode()
-        foreach ($Node in $this.GetDescendentNodes(-1)) {
+        foreach ($Node in $this.GetNodes(-1)) {
             $HashCode = $HashCode -bxor $Node.GetHashCode()
         }
         # Shift the bits to make the level unique
@@ -504,7 +530,7 @@ Class PSMapNode : PSCollectionNode {
 
     [Int]GetHashCode() {
         $HashCode = '@{}'.GetHashCode()
-        foreach ($Node in $this.GetDescendentNodes(-1)) {
+        foreach ($Node in $this.GetNodes(-1)) {
             $HashCode = $HashCode -bxor "$($Node._Name)=$($Node.GetHashCode())".GetHashCode()
         }
         return $HashCode
@@ -540,16 +566,16 @@ Class PSDictionaryNode : PSMapNode {
         $this._Value[$Key] = $Value
     }
 
-    [Collections.Generic.List[PSNode]]GetDescendentNodes([Int]$Levels) {
-        $List = [Collections.Generic.List[PSNode]]::new()
+    [List[PSNode]]GetNodes([Int]$Levels, [PSNodeOrigin]$NodeOrigin, [Bool]$Leaf) {
+        $List = [List[PSNode]]::new()
         if (-not $this.MaxDepthReached()) {
             foreach($Key in $this._Value.get_Keys()) {
                 $Node = $this.Append($this._Value[$Key])
-                $Node._NodeOrigin = 'Map'
                 $Node._Name = $Key
-                $List.Add($Node)
-                if ($Levels -ne 0 -and $Node -is [PSCollectionNode]) {
-                    $list.AddRange($Node.GetDescendentNodes(($Levels - 1)))
+                if ($NodeOrigin -in 0, 'Map' -or ($Leaf -and $Node -is [PSLeafNode])) { $List.Add($Node) }
+                if ($Node -is [PSCollectionNode] -and ($Levels -ne 0 -or $NodeOrigin -eq 'List')) {
+                    $Levels_1 = if ($Levels -gt 0) { $Levels - 1 } else { $Levels }
+                    $list.AddRange($Node.GetNodes($Levels_1, $NodeOrigin, $Leaf))
                 }
             }
         }
@@ -562,7 +588,6 @@ Class PSDictionaryNode : PSMapNode {
             Throw "The <Object>$($this.PathName) doesn't contain a child named: $Key"
         }
         $Node = $this.Append($this._Value[$Key])
-        $Node._NodeOrigin = 'Map'
         $Node._Name = $Key
         return $Node
     }
@@ -597,16 +622,16 @@ Class PSObjectNode : PSMapNode {
         $this._Value.PSObject.Properties[$Name].Value = $Value
     }
 
-    [Collections.Generic.List[PSNode]]GetDescendentNodes([Int]$Levels) {
-        $List = [Collections.Generic.List[PSNode]]::new()
+    [List[PSNode]]GetNodes([Int]$Levels, [PSNodeOrigin]$NodeOrigin, [Bool]$Leaf) {
+        $List = [List[PSNode]]::new()
         if (-not $this.MaxDepthReached()) {
             foreach($Property in $this._Value.PSObject.Properties) {
                 $Node = $this.Append($Property.Value)
-                $Node._NodeOrigin = 'Map'
                 $Node._Name = $Property.Name
-                $List.Add($Node)
-                if ($Levels -ne 0 -and $Node -is [PSCollectionNode]) {
-                    $list.AddRange($Node.GetDescendentNodes($Levels - 1))
+                if ($NodeOrigin -in 0, 'Map' -or ($Leaf -and $Node -is [PSLeafNode])) { $List.Add($Node) }
+                if ($Node -is [PSCollectionNode] -and ($Levels -ne 0 -or $NodeOrigin -eq 'List')) {
+                    $Levels_1 = if ($Levels -gt 0) { $Levels - 1 } else { $Levels }
+                    $list.AddRange($Node.GetNodes($Levels_1, $NodeOrigin, $Leaf))
                 }
             }
         }
@@ -619,7 +644,6 @@ Class PSObjectNode : PSMapNode {
             Throw "The <Object>$($this.PathName) doesn't contain a child named: $Name"
         }
         $Node = $this.Append($this._Value.PSObject.Properties[$Name].Value)
-        $Node._NodeOrigin = 'Map'
         $Node._Name = $Name
         return $Node
     }